remove the infra to accept commands only on certain list
first real tests showed that the race (a msg send to an CC to the
regressions-list might arrive and noticed by regzbot on LKML or other
lists first) is harder to be solved than expected. The functionally was
likely not a good idea from the beginning, so just remove it.