general: rework duplication handling
This makes regzbot commands act local again so they don't affect
all other regressions that are marked as duplicate. But most
will affect the topmost duplicate, to ensure things like "fixes"
propagate there.

This way things are more sane if a duplicate turns out to not be
a duplicate anymore, as the entries than can live again on their
own without being to messy.

Signed-off-by: Thorsten Leemhuis <linux@leemhuis.info>